+++ app/src/main/java/com/battlelancer/seriesguide/shows/episodes/EpisodeDetailsViewModel.kt
@@ -1,5 +1,5 @@
-// Copyright 2023 Uwe Trottmann
 // SPDX-License-Identifier: Apache-2.0
+// Copyright 2021-2025 Uwe Trottmann
 
 package com.battlelancer.seriesguide.shows.episodes
 
@@ -9,19 +9,12 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.switchMap
-import androidx.lifecycle.viewModelScope
 import com.battlelancer.seriesguide.provider.SgRoomDatabase
-import com.battlelancer.seriesguide.streaming.StreamingSearch
 
 class EpisodeDetailsViewModel(
     episodeId: Long,
     application: Application
 ) : AndroidViewModel(application) {
-
-    init {
-        // Set original value for region.
-        StreamingSearch.initRegionLiveData(application)
-    }
 
     val episode = SgRoomDatabase.getInstance(getApplication()).sgEpisode2Helper()
         .getEpisodeLiveData(episodeId)
@@ -29,22 +22,6 @@ class EpisodeDetailsViewModel(
     val show = showId.switchMap {
         SgRoomDatabase.getInstance(getApplication()).sgShow2Helper()
             .getShowLiveData(it)
-    }
-
-    private val showTmdbId = MutableLiveData<Int>()
-    private val watchInfoMediator = StreamingSearch.getWatchInfoMediator(showTmdbId)
-    val watchProvider by lazy {
-        StreamingSearch.getWatchProviderLiveData(
-            watchInfoMediator,
-            viewModelScope.coroutineContext,
-            getApplication()
-        )
-    }
-
-    fun setShowTmdbId(showTmdbId: Int?) {
-        if (showTmdbId != null) {
-            this.showTmdbId.value = showTmdbId
-        }
     }
 
 }