+++ app/src/main/java/com/battlelancer/seriesguide/shows/overview/OverviewFragment.kt
@@ -21,7 +21,9 @@ import androidx.core.view.isInvisible
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
+import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
 import androidx.loader.app.LoaderManager
 import androidx.loader.content.Loader
 import com.battlelancer.seriesguide.R
@@ -52,7 +54,6 @@ import com.battlelancer.seriesguide.shows.tools.ShowStatus
 import com.battlelancer.seriesguide.streaming.StreamingSearch
 import com.battlelancer.seriesguide.streaming.StreamingSearch.initButtons
 import com.battlelancer.seriesguide.tmdbapi.TmdbTools
-import com.battlelancer.seriesguide.tmdbapi.TmdbTools2
 import com.battlelancer.seriesguide.traktapi.CheckInDialogFragment
 import com.battlelancer.seriesguide.traktapi.RateDialogFragment
 import com.battlelancer.seriesguide.traktapi.TraktCredentials
@@ -81,6 +82,7 @@ import com.battlelancer.seriesguide.util.tryStartActivity
 import com.squareup.picasso.Callback
 import com.squareup.picasso.Picasso
 import kotlinx.coroutines.Job
+import kotlinx.coroutines.launch
 import org.greenrobot.eventbus.EventBus
 import org.greenrobot.eventbus.Subscribe
 import org.greenrobot.eventbus.ThreadMode
@@ -236,23 +238,26 @@ class OverviewFragment() : Fragment(), EpisodeActionsContract {
                 sgShow2.nextEpisode.toLong()
             } else -1
             model.setEpisodeId(episodeId)
-            model.setShowTmdbId(sgShow2.tmdbId)
+            sgShow2.tmdbId?.let { model.setShowTmdbId(it) }
         }
         model.episode.observe(viewLifecycleOwner) { sgEpisode2: SgEpisode2? ->
+            // Episode may be null if there is no next episode
             this.binding?.also {
                 maybeAddFeedbackView(it)
-                // May be null if there is no next episode.
                 updateEpisodeViews(it, sgEpisode2)
             }
+            sgEpisode2?.season?.let { model.setSeasonNumber(it) }
         }
-        model.watchProvider.observe(viewLifecycleOwner) { watchInfo: TmdbTools2.WatchInfo? ->
-            if (watchInfo != null) {
-                this.binding?.let {
-                    StreamingSearch.configureButton(
-                        it.includeButtons.buttonEpisodeStreamingSearch,
-                        watchInfo,
-                        requireActivity().getSgAppContainer().preventExternalLinks
-                    )
+        viewLifecycleOwner.lifecycleScope.launch {
+            repeatOnLifecycle(Lifecycle.State.STARTED) {
+                model.seasonWatchInfo.collect { seasonWatchInfo ->
+                    this@OverviewFragment.binding?.let {
+                        StreamingSearch.configureButton(
+                            it.includeButtons.buttonEpisodeStreamingSearch,
+                            seasonWatchInfo,
+                            requireActivity().getSgAppContainer().preventExternalLinks
+                        )
+                    }
                 }
             }
         }