+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Releases marked with 🧪 (or previously with the "beta" suffix) were released o
 * 🔧 Widget: note general item limit in configuration screen, rename "Infinite calendar" option to
   the more applicable "Unlimited days", move hide watch button option to appearance category.
 * 🔧 On Android 16 on Samsung devices, increase the limit of widget items back to 100.
+* 🔧 Shows: only display stream and purchase options specific to the season of a displayed episode.
 
 ## Version 2025.2